refactor: Move unit tests from tests/ to lib.rs
Moves the unit tests from the `tests` directory to the `lib.rs` file for each crate.
This is done to follow the project's new convention of keeping tests in the same file as the code they are testing.

Note: The tests are currently not passing due to an unresolved LLVM dependency issue. This issue needs to be addressed separately.
